Ensure exception caught in jumpToDate is an Error
Prior to this commit, the project failed to build
because the `catch` block of `jumpToDate()` did not
validate the type of the exception before setting
properties on it and forwarding it to methods that
are typed to expect an `Error`.

This commit adds an `instanceof` check to the
`catch` block of `jumpToDate()`.